+++ Changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Exposed missing shutdown handling functions.
 * Fixed incorrect issue url.
+* Generalized app result to `Into<Report>`.
 
 ## [0.1.0] — 2022-05-21